move LoadingSpinner into AsyncContent
